requirements.txt
# Core dependencies
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Document processing
langchain==0.1.0
langchain-community==0.0.13
pypdf==3.17.4
python-docx==1.1.0
openpyxl==3.1.2
beautifulsoup4==4.12.2
lxml==4.9.3

# Text processing
nltk==3.8.1
spacy==3.7.2

# Embeddings
sentence-transformers==2.3.1
torch==2.1.2
transformers==4.36.2

# Vector Database
qdrant-client==1.7.0

# LLM Integration
openai==1.6.1
langchain-openai==0.0.2

# Utilities
requests==2.31.0
tqdm==4.66.1
numpy==1.24.3
pandas==2.1.4

# API
fastapi==0.109.0
uvicorn==0.25.0
pydantic==2.5.0
